+++ examples/with-components/theme/components/Footer.abell
@@ -1,0 +1,10 @@
+<AbellComponent>
+<template>
+  <footer>This is Footer</footer>
+</template>
+<style>
+  footer {
+    color: #f30;
+  }
+</style>
+</AbellComponent>
