+++ examples/with-components/theme/components/Nav.abell
@@ -8,7 +8,7 @@
     <Brand />
   </nav>
 </template>
-<style inlined>
+<style>
   nav {
     background-color: #000;
     color: #fff;
